Reorganize interfaces
