+++ src/interfaces/responses.ts
@@ -1,15 +1,23 @@
 import { BonsaiChapter } from "./bonsai";
+import { UserPartial } from "./user";
 
-// what client receives from backend on get bonsai
-export interface BonsaiPageResponse {
-    geoLocation: string;
-    hardinessZone: string;
-    species: string;
-    height?: string;
-    width?: string;
-    nebari?: string;
-    style?: string;
-    bonsaiChapters: BonsaiChapter[];
-    username: string;
-    id: string;
-  }
+// POST api/auth/login and api/auth/signup
+export interface AuthPostResponse {
+  username?: string;
+  profilePhoto?: string;
+  message?: string;
+}
+
+// GET api/bonsai/:id
+export interface BonsaiGetResponse {
+  geoLocation: string;
+  hardinessZone: string;
+  species: string;
+  height?: string;
+  width?: string;
+  nebari?: string;
+  style?: string;
+  bonsaiChapters: BonsaiChapter[];
+  user: UserPartial;
+  id: string;
+}
